use latest attn rope
